Export a function to register external threads.
Original GC only supports running on the main thread, and
threads that it created itself. When using it for embedded
virtual machines, e.g. Neko, the host application may want
to be able to use it on threads that it created through
system frameworks.

This commit adds a function that can be used to do it.
However, it is a low-level interface that requires knowledge
about the libgc internals to use properly.