Add config option to use STGRTMIN-based signals for thread suspend/resume
Change motivated by using bdwgc in the X10 runtime on systems where
IBM Parallel Environment is used to launch the X10 processes. PE
is already using the signals BDWCG is using for its own purposes, so
we need to force the GC to use different ones to avoid a conflict.