PT_GNU_RELRO: Don't warn about not finding PT_GNU_RELRO segment when we have defined a GC_has_static_roots_func callback. It's most likely not found just because the segment had been excluded.
Alternatively, we could have registered all segments, and checked the callback afterwards, but then we could break
programs that rely on GC_has_static_roots_func to avoid overflowing the maximum number of roots. In addition,
it would make the logic slightly more complicated, probably without a very good reason since the chance of this warning
to show without the segment being excluded is likely to be none.