Workaround 'bad address arithmetic' static analysis tool false positive
The tool complains whether (alloc(size)+ofs) is intentional instead of
(alloc(size+ofs)).  In our case, it is a false alarm (the offset is
added to the result to align the allocation at HBLKSIZE boundary).

* os_dep.c [USE_WINALLOC && MSWIN32] (GC_win32_get_mem): Store result
of GlobalAlloc() to "result" local variable first (then, perform result
alignment in a standalone statement); add comment.