Update shared libraries version info to differentiate against v7.4
There are no API changes in libcord and libgccpp against release-7_4
but some implementation changes exist, 3:*:* version info is reserved
for release-7_4, so the version info for libcord and libgccpp is
now 4:0:3.

There some API changes in libgc against release-7_4, so the "current"
number of the version info is incremented, thus the full version info
is now 4:0:2.

* Makefile.am (LIBGC_VER_INFO): Change version info to 4:0:2
(release-7_4 is supposed to use 3:x:1).
* Makefile.am (LIBGCCPP_VER_INFO): Change version info to 4:0:3
(release-7_4 is supposed to use 3:x:2).
* cord/cord.am (LIBCORD_VER_INFO): Likewise.