Have automake copy instead of symlink, for more git checkout sharability (#8534)
* Make autogen.sh output accurate -- print actual commands run.

* Have automake copy instead of symlink.
 - This sounds analogous to libtool --copy that we do.
 - It enables sharing a source tree more broadly.
   For real example, Cygwin symlinks are not portable to Windows Subsystem For Linux, and probably vice versa.
   For real example, Homebrew symlinks are not portable to Linux in Docker.

Downside I notice is that ls -l is less informative as to what is from git vs. a symlink to outside.
git clean -xxddff works just as well either way.
Hypothetical downside is that updating automake/autoconf doesn't upgrade the copies.
This has never mattered to me, but it might matter. It might be viewed as an advantage.