Update libgc.so version info to differentiate against v8.0.x
There are no API/implementation changes in libcord and libgccpp against
release-8_0 for now, so the corresponding version info is not updated.

There are some implementation changes and new API functions addition in
libgc (without API incompatible changes) against release-8_0, so the
"current" and "age" parts of the version info are incremented.

* Makefile.am (LIBGC_VER_INFO): Change version info to 6:0:5.