Support -fvisibility=hidden option in CMake script
The option is passed only if it is supported and the shared library
build is requested. The functionality is similar to that of configure.

* CMakeLists.txt [BUILD_SHARED_LIBS] (HAVE_FLAG_F_VISIBILITY_HIDDEN):
Check "-fvisibility=hidden" C compiler flag and set the variable
accordingly.
* CMakeLists.txt [BUILD_SHARED_LIBS && HAVE_FLAG_F_VISIBILITY_HIDDEN]
(GC_VISIBILITY_HIDDEN_SET): Define macro.
* CMakeLists.txt [BUILD_SHARED_LIBS && !HAVE_FLAG_F_VISIBILITY_HIDDEN]
(GC_NO_VISIBILITY): Likewise.
* CMakeLists.txt [BUILD_SHARED_LIBS && HAVE_FLAG_F_VISIBILITY_HIDDEN]:
Add "-fvisibility=hidden" compile option.