Use HEURISTIC2 on OpenBSD when single-threaded
Issue #347 (bdwgc).

OpenBSD must use HEURISTIC2 for determining the main thread stack
bottom when threads are disabled due to the random stack gap inserted
by the kernel. In cases where USERSTACK is is visible to user-land, it
is not usable due to the PROT_NONE gap.

* include/private/gcconfig.h [(POWERPC || SPARC || I386 || MIPS
|| HP_PA || ALPHA || AARCH64 || ARM32 || SH && !SH4 || X86_64)
&& OPENBSD && !GC_OPENBSD_THREADS]: Do not include sys/param.h,
uvm/uvm_extern.h.
* include/private/gcconfig.h [(POWERPC || SPARC || I386 || MIPS
|| HP_PA || ALPHA || AARCH64 || ARM32 || SH && !SH4 || X86_64)
&& OPENBSD && !GC_OPENBSD_THREADS && USRSTACK] (STACKBOTTOM): Do not
define.
* include/private/gcconfig.h [(POWERPC || SPARC || I386 || MIPS
|| HP_PA || ALPHA || AARCH64 || ARM32 || SH && !SH4 || X86_64)
&& OPENBSD && !GC_OPENBSD_THREADS && USRSTACK] (HEURISTIC2): Define
macro.