Emscripten support with gctest mostly passing
This adds a bit more in os_dep.c in order to scan the stack and
registers, which is possible if we enable ASYNCIFY (note, this adds
quite some overhead). gctest almost passes with this, except the tests
that involve a_get()

The build system (at least autotools-based) is updated to detect and
configure emscripten correctly. No thread support for now. No idea how
to stop other threads (perhaps we need support from JS side?)

To build

    # source EMSDK first
    emconfigure ./configure
    emmake make gctest.html
    # point your browser at .libs/gctest.html