Fix data race regarding *rlh value in generic_malloc_many
The issue was highlighted in GC_generic_malloc_many() by a static code
analysis tool as "using an unreliable value of *rlh inside the second
locked section; if the data that *rlh depends on was changed by
another thread, this use might be incorrect".

* mallocx.c (GC_generic_malloc_many): Do not increment rlh by lg (use
rlh[lg] instead); change for statement to while; reload rlh value after
acquiring the GC lock (at the end of the loop).