Report if soft-dirty bit is used over mprotect-based implementation
Also, in case of GetWriteWatch-based implementation, do not report
if the feature cannot be used (thus, falling back to mprotect-based
one).

* os_dep.c [GWW_VDB && !SMALL_CONFIG] (detect_GetWriteWatch): Do not
log success and failure cases of the detection.
* os_dep.c [MPROTECT_VDB && (GWW_VDB || SOFT_VDB)] (GC_dirty_init):
Call GC_COND_LOG_PRINTF() with the appropriate message if a call to
GC_gww_dirty_init() or soft_dirty_init() is successful.