Add 'lib' prefix to build artifact names in Makefile.direct
This is to match libraries naming convention of CMake and configure
scripts.

* Makefile.direct: Update comment (add "lib" prefix to gc.a, gccpp.a,
gctba.a, cord.a).
* Makefile.direct (bsd-libgc.a, bsd-libgccpp.a, bsd-libgctba.a,
bsd-libleak.a, base_lib, gc.a, cords, libcord.a, test_cpp, gctba.a,
cord.a, c++, gccpp.a, gctba.a, cordtest, de, gctest): Add "lib" prefix
to gc.a, gccpp.a, gctba.a, cord.a.
* README.QUICK (INSTALLATION, TYPICAL USE): Likewise.
* README.md (Installation and Portability): Likewise.
* doc/gcinterface.md (C/C++ Interface): Likewise.