Check capability tags of derived pointers before pushing to mark stack (a cherry-pick of 90de57c, 256764c, 0c5d8c9 partly from capablevms/bdwgc)
Issue #627 (bdwgc).

Previously, a memory location referenced by an invalid capability had
been eagerly marked as scanned, if it was a plausible collector heap
address.  Now, we no longer mark such a memory location as scanned
since there may be a reachable, valid capability pointing to the same
memory location that we encounter (and therefore should scan) later on
in the mark phase.

Also, ensure that when memory is scanned for capabilities, bounds are
also checked in case a capability with a valid address but invalid
bounds or tags is left in memory.

* include/private/gc_priv.h [!(E2K && USE_PTR_HWTAG) && CHERI_PURECAP]
(LOAD_PTR_OR_CONTINUE): Define as a block statement; define `base_addr`
local variable; `continue` if `cheri_tag_get(v)` is 0 or `v` is out of
capability bounds, or `v` capability do not have `CHERI_PERM_LOAD`.
* include/private/gc_priv.h [CHERI_PURECAP] (CAPABILITY_COVERS_RANGE,
SPANNING_CAPABILITY): Move definition upper.