Check capability tags of derived pointers before pushing to mark-stack.
Previously, a memory location referenced by an invalid capability
had been eagerly marked as scanned, if it was a plausible BDWGC
heap address. Now, we no longer mark such a memory location as scanned
since there may be a reachable, valid capability pointing to the same
memory location that we encounter (and therefore should scan) later on
in the mark phase.