Workaround SIG_SUSPEND delivery to thread inside mutex_lock fail if TSan
Unfortunately, use of an asynchronous signal to suspend threads leads
to the situation when the signal is not delivered (is stored to
pending_signals in ThreadSanitizer runtime actually) while the target
thread is blocked in pthread_mutex_lock.  Thus, we use some synchronous
one instead (which is again unlikely to be used by clients directly).

* include/private/gc_priv.h [SIGNAL_BASED_STOP_WORLD && !SIG_SUSPEND
&& THREAD_SANITIZER] (SIG_SUSPEND): Define to SIGSYS; add comment.
* tests/disclaim.c [GC_PTHREADS] (main): Call
GC_set_thr_restart_signal(GC_get_suspend_signal()) instead of
GC_set_suspend_signal(GC_get_thr_restart_signal()) (thus, the signal
used to suspend threads is not modified in the test).