Workaround SIG_SUSPEND delivery to thread inside mutex_lock fail if TSan (a cherry-pick of commit 0eb8d2f95 from 'master')
Unfortunately, use of an asynchronous signal to suspend threads leads
to the situation when the signal is not delivered (is stored to
pending_signals in ThreadSanitizer runtime actually) while the target
thread is blocked in pthread_mutex_lock.  Thus, we use some synchronous
one instead (which is again unlikely to be used by clients directly).

* include/private/gc_priv.h [GC_PTHREADS && !GC_WIN32_THREADS
&& !NACL && !GC_DARWIN_THREADS && !SIG_SUSPEND && THREAD_SANITIZER]
(SIG_SUSPEND): Define to SIGSYS; add comment.