Utilize stored capability when reporting statistics during free block reclamation
The addresses for the pointers are synthesized and passed into
the GC_reclaim_block() function.  Use capability stored in header
for dereferencing the blocks to be reclaimed.

2. Separate macros for integer and capability aligned access
In debug builds, canary values are written into allocated memory
using load-store macros that assume sizeof(int) = sizeof(int *).
A separate macro is used to calculate the correct offset within
a serialized data structure. This will also prevent data corruption
of other fields within the debug-header.