Fix race between calloc_explicitly_typed and push_complex_descriptor
Issue #449 (bdwgc).

The real problem is that we store the descriptor in the last word of
the object, and then read it in get_complex_descr() as ordinary data,
but we need to make sure that the data in the descriptor is visible to
the GC reading it.  These stores need to be atomic release operations.

* typd_mlc.c (get_complex_descr): Add comment (TODO).
* typd_mlc.c (set_complex_descr): Define macro (using AO_store_release
if available).
* typd_mlc.c (GC_malloc_explicitly_typed, GC_calloc_explicitly_typed,
GC_malloc_explicitly_typed_ignore_off_page): Use set_complex_descr()
instead of direct write to *op.

Co-authored-by: Hans Boehm <boehm@acm.org>