Use capability spanning .data section to scan global roots
Ensure a valid Capability with adequate bounds is utilised to scan
global data area. In the case of CHERI systems where bdwgc is a shared
library, "etext, edata and end" are typeless variables with ezro bounds.
Hence we revert to utilising a capability that spans the entire global
data area returned by dl_iterate_phdr().

Also ensure that when memory is scanned for capabilities, bounds are
also checked in case a capability with a valid address but invalid
bounds or tags is left in memory.