Avoid SIGILL if incremental GC and client is built with optimization (E2K) (fix of commits c5c7c5324, 041cdf23c)
Issue #419 (bdwgc).

In case of e2k, unless -fsemi-spec-ld (or -O0) option is passed to
gcc (both when compiling libgc and the client), a semi-speculative
optimization may lead to SIGILL (instead of SIGSEGV) if mprotect-based
incremental GC is enabled.

* include/private/gcconfig.h [LINUX && !REDIRECT_MALLOC && E2K]
(MPROTECT_VDB): Do not define even if not __OPTIMIZE__; update comment.