Avoid potential race between realloc and GC_block_was_dirty (a cherry-pick of commit ed5281444 from 'master')
GC_realloc might be changing the block size while GC_block_was_dirty
is examining it.  The change to the size field is benign, in that
GC_block_was_dirty would work correctly with either value, since we
are not changing the number of objects in the block.  But seeing
a half-updated value (though unlikely to occur in practice) could be
probably bad.  Using unordered atomic accesses on the size field
should solve the issue.

* mark.c [!GC_DISABLE_INCREMENTAL && AO_HAVE_load]
(GC_block_was_dirty): Use AO_load() to get hhdr->hb_sz value; add
comment.