Allow passing in CFLAGS(_EXTRA)
By using ?= we only set the CFLAGS variable in case it doesn't already
have a value. This means it is now possible for someone to control the
way bdwgc is compiled by feeding in the CFLAGS variable whereas
previously one had to modify the Makefile.direct file. Further, since we
consider the CFLAGS a reasonable default, the primary means of injecting
control flags is by setting CFLAGS_EXTRA, where we now also allow
passing in a value and the stuff in Makefile.direct is appended to that
input.