Allow passing custom CFLAGS[_EXTRA] to Makefile.direct
Issue #483 (bdwgc).

By using ?= we only set the CFLAGS variable in case it doesn't already
have a value.  This means it is now possible for someone to control
the way bdwgc is compiled by feeding in the CFLAGS variable whereas
previously one had to modify the Makefile.direct file.  Further, since
we consider the CFLAGS a reasonable default, the primary means of
injecting control flags is by setting CFLAGS_EXTRA, where we now also
allow passing in a value and the stuff in Makefile.direct is appended
to that input.  Same for [AS_]ABI_FLAG and CXXFLAGS variables.

* Makefile.direct (ABI_FLAG, AS_ABI_FLAG, CFLAGS_EXTRA, CFLAGS,
CXXFLAGS): Assign value using operator "?=" instead of "=" one.