Fix 'unresolved _end' linker error when targeting Android bitcode (a cherry-pick of commit b7676db92 from 'master')
If GC_DONT_REGISTER_MAIN_STATIC_DATA is defined but DYNAMIC_LOADING is
not, GC_register_data_segments will never be called.  However, the
function was still present and referenced symbols not present when
targeting bitcode.  If the linker did not remove the dead function,
linker errors occurred.  Stub an empty function in this case.

* os_dep.c [!OS2 && !MSWIN32 && !MSWINCE && !CYGWIN32 && !AMIGA
&& !OPENBSD && !DYNAMIC_LOADING && GC_DONT_REGISTER_MAIN_STATIC_DATA]
(GC_register_data_segments): Change to no-op (do not use DATASTART and
DATAEND symbols).