Fix store-and-dirty call in GC_CONS (fix of commit 2e7daad62)
GC_PTR_STORE_AND_DIRTY() might be expanded to a debugging version of
the store-and-dirty function, thus the former should not be used for
pointers returned from the non-debugging version of a GC allocation
routine like GC_FAST_MALLOC_GRANS().

* include/gc/gc_inline.h: Add comment that the debugging version of
this API is missing.
* include/gc/gc_inline.h (GC_CONS): Use GC_ptr_store_and_dirty()
instead of GC_PTR_STORE_AND_DIRTY().