Remove EXPORTED_FUNCTIONS linker flag for gctest (Emscripten)
Issue #532 (bdwgc).

When mmap was used, there was apparently a need to export memalign
and memset from the WebAssembly code to JS for the mmap emulation
code to use.

The mmap code path is no longer used with the Emscripten support,
so this export is no longer needed.  The explicit export of _main is
not required; according to the Emscripten code: by default if this
setting is not specified on the command line the _main function will
be implicitly exported.

* tests/tests.am [EMSCRIPTEN] (gctest_html_LDFLAGS): Remove.