Emscripten, gctest: Remove EXPORTED_FUNCTIONS.
When `mmap` was used, there was apparently a need to export
`memalign` and `memset` from the WebAssembly code to JS for
the mmap emulation code to use.

The `mmap` code path is no longer used with the Emscripten support,
so this export is no longer needed.

The explicit export of `_main` is not required. According to the
Emscripten code:

    By default if this setting is not specified on the command
    line the `_main` function will be implicitly exported.

* tests/tests.am [EMSCRIPTEN]: Remove `gctest_html_LDFLAGS`.