Fix description of client promise for IGNORE_OFF_PAGE allocated objects (a cherry-pick of commit d1aa5274a from 'release-8_2')
The client should keep a pointer within the first heap block for such
an allocated object.  Previously, it was mentioned in the documentation
and comments that such a pointer should be within the first 256 or 512
bytes.

* README.md (The C Interface to the Allocator): Fix the description
(comment) of GC_malloc_ignore_off_page (the client promises to keep
a pointer within the first hblk of the object instead of 256 or 512
bytes, or a page).
* doc/gcinterface.md (GC_MALLOC_IGNORE_OFF_PAGE): Likewise.
* doc/gc.man (GC_malloc_atomic_ignore_off_page): Likewise.
* include/gc.h (GC_malloc_ignore_off_page): Likewise.
* include/gc_mark.h (GC_generic_malloc_ignore_off_page): Likewise.
* include/private/gc_priv.h (IGNORE_OFF_PAGE): Likewise.
* include/private/gc_priv.h [DBG_HDRS_ALL || GC_GCJ_SUPPORT
|| !GC_NO_FINALIZATION] (GC_generic_malloc_inner_ignore_off_page):
Likewise.
* malloc.c [DBG_HDRS_ALL || GC_GCJ_SUPPORT || !GC_NO_FINALIZATION]
(GC_generic_malloc_inner_ignore_off_page): Likewise.
* include/gc_gcj.h (GC_gcj_malloc_ignore_off_page): Refine comment.