Fix signals delivery fail in find-leak mode if init from non-main thread (a cherry-pick of commits f806e6209, 68d419a51 from 'master')
Issue #542 (bdwgc).

The scenario of the failure (on Linux):
- Find-leak mode is on
- pthread_create/join primitives are not redirected
- GC is initialized from a thread other than primordial
- that thread is terminated (but not unregistered as the 1st one in GC)
- the primordial thread exits and launches GC_exit_check
- GC_gcollect tries to suspend the terminated thread by sending a signal
- the signal is not delivered but pthread_kill just returns zero
- the signal sending is retried until aborting with the proper message.

The proposed solution (workaround) is not to call GC_gcollect at
the process exit (in the find-leak mode) if GC_init was called from
a thread other than primordial one.

Check in GC_exit_check() (which is called if GC_find_leak) that the
thread executing at-exit functions is the same as the one performed
the GC initialization, otherwise the latter thread might already be
dead but still registered and this, as a consequence, might cause
a signal delivery fail when suspending the threads on platforms that
do not guarantee ESRCH returned if the signal is not delivered (as
observed on Ubuntu 22).

* include/private/gc_priv.h [THREADS && !DONT_USE_ATEXIT]
(GC_is_main_thread): Declare GC_INNER function.
* misc.c [!DONT_USE_ATEXIT && THREADS] (GC_exit_check): If not
GC_is_main_thread() or not GC_thread_is_registered() then do not call
GC_gcollect(); add comment; do not set and clear GC_in_thread_creation.
* pthread_support.c [!DONT_USE_ATEXIT] (GC_main_thread_id): Define
STATIC variable.
* pthread_support.c [!DONT_USE_ATEXIT] (GC_is_main_thread): Implement.
* win32_threads.c [!DONT_USE_ATEXIT] (GC_is_main_thread): Likewise.
* pthread_support.c [!DONT_USE_ATEXIT] (GC_thr_init): Set
GC_main_thread_id.
* win32_threads.c (GC_main_thread): Define variable unconditionally;
do not initialize to zero.
* win32_threads.c [!DONT_USE_ATEXIT] (GC_thr_init): Set
GC_main_thread variable unconditionally; do not undefine
GC_main_thread.