Add support for building with Zig
This adds a build.zig that allows building the garbage collector using
the Zig build system. It implements a subset of the configuration
options offered by CMake and following the naming convention in CMake
for all those options.

There are two new GitHub Actions workflows:
- zig build is essentially a copy of the cmake build workflow
  - it builds the collector and runs tests
- zig xbuild uses zig to cross-compile for different platforms
  - we don't actually run the tests as that would require emulation

The README is slightly updated, not just adding an explanation of zig
but restructuring the whole section on how to build the GC somewhat.

The build.zig follows the format used in zig v0.12. The final 0.12
version is not out yet, so users will have to use a build from the zig
master branch for now. This is not ideal but the whole zig build system
is so new that there have been backwards incompatible changes from 0.11
which is why we opt to use 0.12 that is likely going to be closer to the
final form.

There is no support for installing documentation currently.