Prevent hang if a thread acquiring reader lock is suspended externally (fix of commit 3bfb499f3)
Issue #473 (bdwgc).

At least in the Linux threads implementation, rwlock primitives are not
atomic in respect to signals, and suspending externally a thread which
is running inside pthread_rwlock_rdlock() may lead to a deadlock.

As a workaround GC_suspend_thread() API is disabled if USE_RWLOCK.

* include/private/gcconfig.h [USE_RWLOCK] (GC_ENABLE_SUSPEND_THREAD):
Undefine; add comment.